src/test/java/com/es/phoneshop/web/MiniCartServletTest.java
package com.es.phoneshop.web;

import com.es.phoneshop.cart.Cart;
import com.es.phoneshop.cart.CartService;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.MockitoJUnitRunner;

import javax.servlet.RequestDispatcher;
import javax.servlet.ServletException;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

import static org.mockito.ArgumentMatchers.anyString;
import static org.mockito.ArgumentMatchers.eq;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

@RunWith(MockitoJUnitRunner.class)
public class MiniCartServletTest {
    @Mock
    private HttpServletRequest request;
    @Mock
    private HttpServletResponse response;
    @Mock
    private RequestDispatcher requestDispatcher;
    @Mock
    private CartService cartService;
    @Mock
    private Cart cart;
    @InjectMocks
    private MiniCartServlet miniCartServlet;

    @Before
    public void setup() {
        when(cartService.getCart(request.getSession())).thenReturn(cart);
        when(request.getRequestDispatcher(anyString())).thenReturn(requestDispatcher);
    }

    @Test
    public void doGet() throws ServletException, IOException {
        miniCartServlet.doGet(request, response);

        verify(request).setAttribute(eq("cart"), eq(cart));
    }

    @Test
    public void doPost() throws ServletException, IOException {
        miniCartServlet.doPost(request, response);

        verify(request).setAttribute(eq("cart"), eq(cart));
    }
}
